+++ crates/common/src/lib.rs
@@ -19,8 +19,9 @@ pub mod telemetry;
 pub mod worker;
 pub use temporalio_common_wasm::{
     ActivityDefinition, ActivityError, HasWorkflowDefinition, Priority, QueryDefinition,
-    SignalDefinition, UntypedActivity, UntypedWorkflow, UpdateDefinition, WorkerDeploymentVersion,
-    WorkflowDefinition, WorkflowExecution, data_converters, error, search_attributes,
+    RetryPolicy, SignalDefinition, UntypedActivity, UntypedWorkflow, UpdateDefinition,
+    WorkerDeploymentVersion, WorkflowDefinition, WorkflowExecution, data_converters, error,
+    search_attributes,
 };
 
 macro_rules! dbg_panic {