+++ crates/workflow/src/lib.rs
@@ -26,9 +26,12 @@ pub use runtime::model::{CancellableID, UnblockEvent};
 pub use runtime::model::{TimerResult, WorkflowResult, WorkflowTermination};
 #[doc(hidden)]
 pub use runtime::{SdkWakeGuard, is_sdk_wake};
-pub use temporalio_common_wasm::error::{
-    ActivityExecutionError, ChildWorkflowExecutionError, ChildWorkflowStartError, RetryState,
-    TimeoutType, WorkflowSignalError,
+pub use temporalio_common_wasm::{
+    RetryPolicy,
+    error::{
+        ActivityExecutionError, ChildWorkflowExecutionError, ChildWorkflowStartError, RetryState,
+        TimeoutType, WorkflowSignalError,
+    },
 };
 pub use workflow_context::{
     ActivityCancellationType, ActivityCloseTimeouts, ActivityOptions, BaseWorkflowContext,